Extracting common baseclass for L2B/C processes capable of async calls. Switching to iterative mechanism when extracting data from L2B/C files.
